feat: move tabs and components to specific folder
