+++ src/components/General/RightSideBar/ValidatorsTabs/ValidatorsTabs.tsx
